requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
crispy-bootstrap5==2024.2
Django==4.2.14
django-crispy-forms==2.2
python-dotenv==1.0.1
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
